+++ sdk/inc/azure/iot/az_iot_common.h
@@ -266,30 +266,6 @@ AZ_NODISCARD AZ_INLINE bool az_iot_status_retriable(az_iot_status status)
   return ((status == AZ_IOT_STATUS_THROTTLED) || (status == AZ_IOT_STATUS_SERVER_ERROR));
 }
 
-/**
- * @brief Calculates the recommended delay before retrying an operation that failed.
- *
- * @param[in] operation_msec The time it took, in milliseconds, to perform the operation that
- *                           failed.
- * @param[in] attempt The number of failed retry attempts.
- * @param[in] min_retry_delay_msec The minimum time, in milliseconds, to wait before a retry.
- * @param[in] max_retry_delay_msec The maximum time, in milliseconds, to wait before a retry.
- * @param[in] random_jitter_msec A random value between 0 and the maximum allowed jitter, in
- * milliseconds.
- * @pre \p operation_msec must be between 0 and INT32_MAX - 1.
- * @pre \p attempt must be between 0 and INT16_MAX - 1.
- * @pre \p min_retry_delay_msec must be between 0 and INT32_MAX - 1.
- * @pre \p max_retry_delay_msec must be between 0 and INT32_MAX - 1.
- * @pre \p random_jitter_msec must be between 0 and INT32_MAX - 1.
- * @return The recommended delay in milliseconds.
- */
-AZ_NODISCARD int32_t az_iot_calculate_retry_delay(
-    int32_t operation_msec,
-    int16_t attempt,
-    int32_t min_retry_delay_msec,
-    int32_t max_retry_delay_msec,
-    int32_t random_jitter_msec);
-
 #include <azure/core/_az_cfg_suffix.h>
 
 #endif // _az_IOT_CORE_H
