+++ sdk/src/azure/core/CMakeLists.txt
@@ -71,6 +71,7 @@ if (AZ_PLATFORM_IMPL STREQUAL "POSIX" AND NOT APPLE)
     ${CMAKE_CURRENT_LIST_DIR}/az_mqtt5_request_hfsm.c
     ${CMAKE_CURRENT_LIST_DIR}/az_mqtt5_topic_parser.c
     ${CMAKE_CURRENT_LIST_DIR}/az_precondition.c
+    ${CMAKE_CURRENT_LIST_DIR}/az_retry.c
     ${CMAKE_CURRENT_LIST_DIR}/az_span.c
   )
 else()
