+++ sdk/tests/iot/common/test_az_iot_common.c
@@ -268,87 +268,6 @@ static void test_az_iot_status_retriable_translate_success()
   assert_false(az_iot_status_retriable(AZ_IOT_STATUS_UNAUTHORIZED));
 }
 
-static void test_az_iot_calculate_retry_delay_common_timings_success()
-{
-  assert_int_equal(2229, az_iot_calculate_retry_delay(5, 1, 500, 100000, 1234));
-  assert_int_equal(321, az_iot_calculate_retry_delay(5000, 1, 500, 100000, 4321));
-
-  // Operation already took more than the back-off interval.
-  assert_int_equal(0, az_iot_calculate_retry_delay(10000, 1, 500, 100000, 4321));
-
-  // Max retry exceeded.
-  assert_int_equal(9995, az_iot_calculate_retry_delay(5, 5, 500, 10000, 4321));
-}
-
-static void test_az_iot_calculate_retry_delay_overflow_time_success()
-{
-  assert_int_equal(
-      0,
-      az_iot_calculate_retry_delay(
-          INT32_MAX - 1, INT16_MAX - 1, INT32_MAX - 1, INT32_MAX - 1, INT32_MAX - 1));
-
-  assert_int_equal(
-      INT32_MAX - 1,
-      az_iot_calculate_retry_delay(0, INT16_MAX - 1, INT32_MAX - 1, INT32_MAX - 1, INT32_MAX - 1));
-}
-
-static int _log_retry = 0;
-static void _log_listener(az_log_classification classification, az_span message)
-{
-  switch (classification)
-  {
-    case AZ_LOG_IOT_RETRY:
-      _log_retry++;
-      assert_int_equal(az_span_size(message), 0);
-      break;
-    default:
-      assert_true(false);
-  }
-}
-
-static bool _should_write_iot_retry_only(az_log_classification classification)
-{
-  switch (classification)
-  {
-    case AZ_LOG_IOT_RETRY:
-      return true;
-    default:
-      return false;
-  }
-}
-
-static bool _should_write_nothing(az_log_classification classification)
-{
-  (void)classification;
-  return false;
-}
-
-static void test_az_iot_calculate_retry_delay_logging_succeed()
-{
-  az_log_set_message_callback(_log_listener);
-  az_log_set_classification_filter_callback(_should_write_iot_retry_only);
-
-  _log_retry = 0;
-  assert_int_equal(2229, az_iot_calculate_retry_delay(5, 1, 500, 100000, 1234));
-  assert_int_equal(_az_BUILT_WITH_LOGGING(1, 0), _log_retry);
-
-  az_log_set_message_callback(NULL);
-  az_log_set_classification_filter_callback(NULL);
-}
-
-static void test_az_iot_calculate_retry_delay_no_logging_succeed()
-{
-  az_log_set_message_callback(_log_listener);
-  az_log_set_classification_filter_callback(_should_write_nothing);
-
-  _log_retry = 0;
-  assert_int_equal(2229, az_iot_calculate_retry_delay(5, 1, 500, 100000, 1234));
-  assert_int_equal(_az_BUILT_WITH_LOGGING(0, 0), _log_retry);
-
-  az_log_set_message_callback(NULL);
-  az_log_set_classification_filter_callback(NULL);
-}
-
 static void test_az_span_copy_url_encode_succeed()
 {
   az_span url_decoded_span = AZ_SPAN_FROM_STR("abc/=%012");
@@ -795,10 +714,6 @@ int test_az_iot_common()
     cmocka_unit_test(test_az_iot_u64toa_size_success),
     cmocka_unit_test(test_az_iot_is_status_succeeded_translate_success),
     cmocka_unit_test(test_az_iot_status_retriable_translate_success),
-    cmocka_unit_test(test_az_iot_calculate_retry_delay_common_timings_success),
-    cmocka_unit_test(test_az_iot_calculate_retry_delay_overflow_time_success),
-    cmocka_unit_test(test_az_iot_calculate_retry_delay_logging_succeed),
-    cmocka_unit_test(test_az_iot_calculate_retry_delay_no_logging_succeed),
     cmocka_unit_test(test_az_span_copy_url_encode_succeed),
     cmocka_unit_test(test_az_span_copy_url_encode_insufficient_size_fail),
     cmocka_unit_test(test_az_iot_message_properties_init_succeed),
